[move] #6 TheaterClassificationTabInModal.kt 생성 후 이곳으로 TheaterClassificationTabInModal 함수 이동
