+++ app/src/main/java/org/sopt/cgv/feature/time/TheaterClassificationTabInModal.kt
@@ -1,0 +1,58 @@
+package org.sopt.cgv.feature.time
+
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Tab
+import androidx.compose.material3.TabRow
+import androidx.compose.material3.TabRowDefaults
+import androidx.compose.material3.TabRowDefaults.tabIndicatorOffset
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.MutableState
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.unit.dp
+import org.sopt.cgv.core.designsystem.theme.CGVTheme
+
+@Composable
+fun TheaterClassificationTabInModal(
+    selectedIndex: MutableState<Int>,
+    tabs: List<String>
+) {
+    Column(
+        modifier = Modifier
+            .fillMaxWidth()
+            .padding(horizontal = 18.dp)
+    ) {
+        TabRow(
+            selectedTabIndex = selectedIndex.value,
+            modifier = Modifier,
+            containerColor = Color.White,
+            contentColor = Color.White,
+            indicator = { tabPositions ->
+                TabRowDefaults.PrimaryIndicator(
+                    modifier = Modifier.tabIndicatorOffset(tabPositions[selectedIndex.value]),
+                    width = 175.dp,
+                    color = Color.Red
+                )
+            },
+            divider = {}
+        ) {
+            tabs.forEachIndexed { index, category ->
+                Tab(
+                    selected = selectedIndex.value == index,
+                    onClick = { selectedIndex.value = index },
+                    modifier = Modifier.padding(vertical = 10.dp),
+                    selectedContentColor = Color.Red,
+                    unselectedContentColor = Color.Black
+                ) {
+                    Text(
+                        text = category,
+                        style = CGVTheme.typography.body3_m_14
+                    )
+                }
+            }
+        }
+    }
+}
