+++ app/src/main/java/org/sopt/cgv/feature/time/TimeScreen.kt
@@ -2,26 +2,15 @@ package org.sopt.cgv.feature.time
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Button
 import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.Tab
-import androidx.compose.material3.TabRow
-import androidx.compose.material3.TabRowDefaults
-import androidx.compose.material3.TabRowDefaults.tabIndicatorOffset
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.MutableState
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import org.sopt.cgv.core.designsystem.theme.CGVTheme
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
@@ -42,48 +31,6 @@ fun TimeScreen(
         isSheetOpen = isSheetOpen,
         onDismissRequest = { isSheetOpen = false }
     )
-}
-
-@Composable
-fun TheaterClassificationTabInModal(
-    selectedIndex: MutableState<Int>,
-    tabs: List<String>
-) {
-    Column(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(horizontal = 18.dp)
-    ) {
-        TabRow(
-            selectedTabIndex = selectedIndex.value,
-            modifier = Modifier,
-            containerColor = Color.White,
-            contentColor = Color.White,
-            indicator = { tabPositions ->
-                TabRowDefaults.PrimaryIndicator(
-                    modifier = Modifier.tabIndicatorOffset(tabPositions[selectedIndex.value]),
-                    width = 175.dp,
-                    color = Color.Red
-                )
-            },
-            divider = {}
-        ) {
-            tabs.forEachIndexed { index, category ->
-                Tab(
-                    selected = selectedIndex.value == index,
-                    onClick = { selectedIndex.value = index },
-                    modifier = Modifier.padding(vertical = 10.dp),
-                    selectedContentColor = Color.Red,
-                    unselectedContentColor = Color.Black
-                ) {
-                    Text(
-                        text = category,
-                        style = CGVTheme.typography.body3_m_14
-                    )
-                }
-            }
-        }
-    }
 }
 
 @Preview
